pyproject.toml
[tool.poetry]
name = "fastapi-mcp-demo"
version = "0.1.0"
description = ""
authors = ["jggomez <juan.gomez01@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi-mcp = "^0.3.0"
fastapi = "^0.115.12"
mcp = {extras = ["cli"], version = "^1.6.0"}
google-genai = "^1.11.0"
mcp-proxy = "^0.5.1"
rich = "^14.0.0"
starlette = "^0.46.2"
python-dotenv = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
